+++ src/Model/CubismMoc.cpp
@@ -10,12 +10,26 @@
 
 namespace Live2D { namespace Cubism { namespace Framework {
 
-CubismMoc* CubismMoc::Create(const csmByte* mocBytes, csmSizeInt size)
+CubismMoc* CubismMoc::Create(const csmByte* mocBytes, csmSizeInt size, csmBool shouldCheckMocConsistency)
 {
     CubismMoc* cubismMoc = NULL;
 
     void* alignedBuffer = CSM_MALLOC_ALLIGNED(size, Core::csmAlignofMoc);
     memcpy(alignedBuffer, mocBytes, size);
+
+    if (shouldCheckMocConsistency)
+    {
+        // .moc3の整合性を確認
+        csmBool consistency = HasMocConsistency(alignedBuffer, size);
+        if (!consistency)
+        {
+            CSM_FREE_ALLIGNED(alignedBuffer);
+
+            // 整合性が確認できなければ処理しない
+            CubismLogError("Inconsistent MOC3.");
+            return cubismMoc;
+        }
+    }
 
     Core::csmMoc* moc = Core::csmReviveMocInPlace(alignedBuffer, size);
     const Core::csmMocVersion version = Core::csmGetMocVersion(alignedBuffer, size);
@@ -84,8 +98,20 @@ Core::csmMocVersion CubismMoc::GetMocVersion()
 
 csmBool CubismMoc::HasMocConsistency(void* address, const csmUint32 size)
 {
-    csmInt32 consistencyFlags = Core::csmHasMocConsistency(address, size);
-    return consistencyFlags != 0 ? true : false;
+    csmInt32 isConsistent = Core::csmHasMocConsistency(address, size);
+    return isConsistent != 0 ? true : false;
+}
+
+csmBool CubismMoc::HasMocConsistencyFromUnrevivedMoc(const csmByte* mocBytes, csmSizeInt size)
+{
+    void* alignedBuffer = CSM_MALLOC_ALLIGNED(size, Core::csmAlignofMoc);
+    memcpy(alignedBuffer, mocBytes, size);
+
+    csmBool consistency = CubismMoc::HasMocConsistency(alignedBuffer, size);
+
+    CSM_FREE_ALLIGNED(alignedBuffer);
+
+    return consistency;
 }
 
 }}}