+++ src/Model/CubismModel.cpp
@@ -26,6 +26,7 @@ CubismModel::CubismModel(Core::csmModel* model)
     , _isOverwrittenModelMultiplyColors(false)
     , _isOverwrittenModelScreenColors(false)
     , _isOverwrittenCullings(false)
+    , _modelOpacity(1.0f)
 { }
 
 CubismModel::~CubismModel()
@@ -378,15 +379,19 @@ void CubismModel::Initialize()
         }
     }
 
+    const csmInt32  partCount = Core::csmGetPartCount(_model);
     {
         const csmChar** partIds = Core::csmGetPartIds(_model);
-        const csmInt32  partCount = Core::csmGetPartCount(_model);
 
         _partIds.PrepareCapacity(partCount);
         for (csmInt32 i = 0; i < partCount; ++i)
         {
             _partIds.PushBack(CubismFramework::GetIdManager()->GetId(partIds[i]));
         }
+
+        _userPartMultiplyColors.PrepareCapacity(partCount);
+        _userPartScreenColors.PrepareCapacity(partCount);
+        _partChildDrawables.Resize(partCount);
     }
 
     {
@@ -398,15 +403,17 @@ void CubismModel::Initialize()
         _userScreenColors.PrepareCapacity(drawableCount);
         _userCullings.PrepareCapacity(drawableCount);
 
+        // カリング設定
+        DrawableCullingData userCulling;
+        userCulling.IsOverwritten = false;
+        userCulling.IsCulling = 0;
+
         // 乗算色
         Rendering::CubismRenderer::CubismTextureColor multiplyColor;
         multiplyColor.R = 1.0f;
         multiplyColor.G = 1.0f;
         multiplyColor.B = 1.0f;
         multiplyColor.A = 1.0f;
-        DrawableColorData userMultiplyColor;
-        userMultiplyColor.IsOverwritten = false;
-        userMultiplyColor.Color = multiplyColor;
 
         // スクリーン色
         Rendering::CubismRenderer::CubismTextureColor screenColor;
@@ -414,22 +421,51 @@ void CubismModel::Initialize()
         screenColor.G = 0.0f;
         screenColor.B = 0.0f;
         screenColor.A = 1.0f;
-        DrawableColorData userScreenColor;
-        userScreenColor.IsOverwritten = false;
-        userScreenColor.Color = screenColor;
 
-        // カリング設定
-        DrawableCullingData userCulling;
-        userCulling.IsOverwritten = false;
-        userCulling.IsCulling = 0;
-
-
-        for (csmInt32 i = 0; i < drawableCount; ++i)
+        // Parts
         {
-            _drawableIds.PushBack(CubismFramework::GetIdManager()->GetId(drawableIds[i]));
-            _userMultiplyColors.PushBack(userMultiplyColor);
-            _userScreenColors.PushBack(userScreenColor);
-            _userCullings.PushBack(userCulling);
+            // 乗算色
+            PartColorData userMultiplyColor;
+            userMultiplyColor.IsOverwritten = false;
+            userMultiplyColor.Color = multiplyColor;
+
+            // スクリーン色
+            PartColorData userScreenColor;
+            userScreenColor.IsOverwritten = false;
+            userScreenColor.Color = screenColor;
+
+            for (csmInt32 i = 0; i < partCount; ++i)
+            {
+                _userPartMultiplyColors.PushBack(userMultiplyColor);
+                _userPartScreenColors.PushBack(userScreenColor);
+            }
+        }
+
+        // Drawables
+        {
+            // 乗算色
+            DrawableColorData userMultiplyColor;
+            userMultiplyColor.IsOverwritten = false;
+            userMultiplyColor.Color = multiplyColor;
+
+            // スクリーン色
+            DrawableColorData userScreenColor;
+            userScreenColor.IsOverwritten = false;
+            userScreenColor.Color = screenColor;
+
+            for (csmInt32 i = 0; i < drawableCount; ++i)
+            {
+                _drawableIds.PushBack(CubismFramework::GetIdManager()->GetId(drawableIds[i]));
+                _userMultiplyColors.PushBack(userMultiplyColor);
+                _userScreenColors.PushBack(userScreenColor);
+                _userCullings.PushBack(userCulling);
+
+                csmInt32 parentIndex = Core::csmGetDrawableParentPartIndices(_model)[i];
+                if (parentIndex >= 0)
+                {
+                    _partChildDrawables[parentIndex].PushBack(i);
+                }
+            }
         }
     }
 }
@@ -438,6 +474,12 @@ CubismIdHandle CubismModel::GetDrawableId(csmInt32 drawableIndex) const
 {
     const csmChar** parameterIds = Core::csmGetDrawableIds(_model);
     return CubismFramework::GetIdManager()->GetId(parameterIds[drawableIndex]);
+}
+
+CubismIdHandle CubismModel::GetPartId(csmUint32 partIndex)
+{
+    const csmChar** partIds = Core::csmGetPartIds(_model);
+    return CubismFramework::GetIdManager()->GetId(partIds[partIndex]);
 }
 
 csmInt32 CubismModel::GetPartCount() const
@@ -675,6 +717,60 @@ void CubismModel::SetScreenColor(csmInt32 drawableIndex, csmFloat32 r, csmFloat3
     _userScreenColors[drawableIndex].Color.A = a;
 }
 
+Rendering::CubismRenderer::CubismTextureColor CubismModel::GetPartMultiplyColor(csmInt32 partIndex) const
+{
+    return _userPartMultiplyColors[partIndex].Color;
+}
+
+Rendering::CubismRenderer::CubismTextureColor CubismModel::GetPartScreenColor(csmInt32 partIndex) const
+{
+    return _userPartScreenColors[partIndex].Color;
+}
+
+void CubismModel::SetPartMultiplyColor(csmInt32 partIndex, const Rendering::CubismRenderer::CubismTextureColor& color)
+{
+    SetPartMultiplyColor(partIndex, color.R, color.G, color.B, color.A);
+}
+
+void CubismModel::SetPartColor(
+    csmUint32 partIndex,
+    csmFloat32 r, csmFloat32 g, csmFloat32 b, csmFloat32 a,
+    csmVector<PartColorData>& partColors,
+    csmVector <DrawableColorData>& drawableColors)
+{
+    partColors[partIndex].Color.R = r;
+    partColors[partIndex].Color.G = g;
+    partColors[partIndex].Color.B = b;
+    partColors[partIndex].Color.A = a;
+
+    if (partColors[partIndex].IsOverwritten)
+    {
+        for (csmUint32 i = 0; i < _partChildDrawables[partIndex].GetSize(); i++)
+        {
+            csmUint32 drawableIndex = _partChildDrawables[partIndex][i];
+            drawableColors[drawableIndex].Color.R = r;
+            drawableColors[drawableIndex].Color.G = g;
+            drawableColors[drawableIndex].Color.B = b;
+            drawableColors[drawableIndex].Color.A = a;
+        }
+    }
+}
+
+void CubismModel::SetPartMultiplyColor(csmInt32 partIndex, csmFloat32 r, csmFloat32 g, csmFloat32 b, csmFloat32 a)
+{
+    SetPartColor(partIndex, r, g, b, a, _userPartMultiplyColors, _userMultiplyColors);
+}
+
+void CubismModel::SetPartScreenColor(csmInt32 partIndex, const Rendering::CubismRenderer::CubismTextureColor& color)
+{
+    SetPartScreenColor(partIndex, color.R, color.G, color.B, color.A);
+}
+
+void CubismModel::SetPartScreenColor(csmInt32 partIndex, csmFloat32 r, csmFloat32 g, csmFloat32 b, csmFloat32 a)
+{
+    SetPartColor(partIndex, r, g, b, a, _userPartScreenColors, _userScreenColors);
+}
+
 csmBool CubismModel::GetOverwriteFlagForModelMultiplyColors() const
 {
     return _isOverwrittenModelMultiplyColors;
@@ -715,6 +811,50 @@ void CubismModel::SetOverwriteFlagForDrawableScreenColors(csmUint32 drawableInde
     _userScreenColors[drawableIndex].IsOverwritten = value;
 }
 
+csmBool CubismModel::GetOverwriteColorForPartMultiplyColors(csmInt32 partIndex) const
+{
+    return _userPartMultiplyColors[partIndex].IsOverwritten;
+}
+
+csmBool CubismModel::GetOverwriteColorForPartScreenColors(csmInt32 partIndex) const
+{
+    return _userPartScreenColors[partIndex].IsOverwritten;
+}
+
+void CubismModel::SetOverwriteColorForPartColors(
+    csmUint32 partIndex,
+    csmBool value,
+    csmVector<PartColorData>& partColors,
+    csmVector <DrawableColorData>& drawableColors)
+{
+    partColors[partIndex].IsOverwritten = value;
+
+    for (csmUint32 i = 0; i < _partChildDrawables[partIndex].GetSize(); i++)
+    {
+        csmUint32 drawableIndex = _partChildDrawables[partIndex][i];
+        drawableColors[drawableIndex].IsOverwritten = value;
+        if (value)
+        {
+            drawableColors[drawableIndex].Color.R = partColors[partIndex].Color.R;
+            drawableColors[drawableIndex].Color.G = partColors[partIndex].Color.G;
+            drawableColors[drawableIndex].Color.B = partColors[partIndex].Color.B;
+            drawableColors[drawableIndex].Color.A = partColors[partIndex].Color.A;
+        }
+    }
+}
+
+void CubismModel::SetOverwriteColorForPartMultiplyColors(csmUint32 partIndex, csmBool value)
+{
+    _userPartMultiplyColors[partIndex].IsOverwritten = value;
+    SetOverwriteColorForPartColors(partIndex, value, _userPartMultiplyColors, _userMultiplyColors);
+}
+
+void CubismModel::SetOverwriteColorForPartScreenColors(csmUint32 partIndex, csmBool value)
+{
+    _userPartScreenColors[partIndex].IsOverwritten = value;
+    SetOverwriteColorForPartColors(partIndex, value, _userPartScreenColors, _userScreenColors);
+}
+
 csmInt32 CubismModel::GetDrawableCulling(csmInt32 drawableIndex) const
 {
     if (GetOverwriteFlagForModelCullings() || GetOverwriteFlagForDrawableCullings(drawableIndex))
@@ -749,6 +889,16 @@ csmBool CubismModel::GetOverwriteFlagForDrawableCullings(csmInt32 drawableIndex)
 void CubismModel::SetOverwriteFlagForDrawableCullings(csmUint32 drawableIndex, csmBool value)
 {
     _userCullings[drawableIndex].IsOverwritten = value;
+}
+
+csmFloat32 CubismModel::GetModelOpacity()
+{
+    return _modelOpacity;
+}
+
+void CubismModel::SetModelOpacity(csmFloat32 value)
+{
+    _modelOpacity = value;
 }
 
 Core::csmModel* CubismModel::GetModel() const