+++ src/Motion/CubismExpressionMotion.cpp
@@ -33,22 +33,54 @@ CubismExpressionMotion::~CubismExpressionMotion()
 CubismExpressionMotion* CubismExpressionMotion::Create(const csmByte* buffer, csmSizeInt size)
 {
     CubismExpressionMotion* expression = CSM_NEW CubismExpressionMotion();
+    expression->Parse(buffer, size);
+    return expression;
+}
 
+void CubismExpressionMotion::DoUpdateParameters(CubismModel* model, csmFloat32 userTimeSeconds, csmFloat32 weight, CubismMotionQueueEntry* motionQueueEntry)
+{
+    for (csmUint32 i = 0; i < _parameters.GetSize(); ++i)
+    {
+        ExpressionParameter& parameter = _parameters[i];
+
+        switch (parameter.BlendType)
+        {
+        case ExpressionBlendType_Add: {
+            model->AddParameterValue(parameter.ParameterId, parameter.Value, weight);            // 相対変化 加算
+            break;
+        }
+        case ExpressionBlendType_Multiply: {
+            model->MultiplyParameterValue(parameter.ParameterId, parameter.Value, weight);       // 相対変化 乗算
+            break;
+        }
+        case ExpressionBlendType_Overwrite: {
+            model->SetParameterValue(parameter.ParameterId, parameter.Value, weight);            // 絶対変化 上書き
+            break;
+        }
+        default:
+            // 仕様にない値を設定したときは既に加算モードになっている
+            break;
+        }
+    }
+}
+
+void CubismExpressionMotion::Parse(const csmByte* buffer, csmSizeInt size)
+{
     Utils::CubismJson* json = Utils::CubismJson::Create(buffer, size);
     Utils::Value& root = json->GetRoot();
 
-    expression->SetFadeInTime(root[ExpressionKeyFadeIn].ToFloat(DefaultFadeTime));   // フェードイン
-    expression->SetFadeOutTime(root[ExpressionKeyFadeOut].ToFloat(DefaultFadeTime)); // フェードアウト
+    SetFadeInTime(root[ExpressionKeyFadeIn].ToFloat(DefaultFadeTime));   // フェードイン
+    SetFadeOutTime(root[ExpressionKeyFadeOut].ToFloat(DefaultFadeTime)); // フェードアウト
 
     // 各パラメータについて
     const csmInt32 parameterCount = root[ExpressionKeyParameters].GetSize();
-    expression->_parameters.PrepareCapacity(parameterCount);
+    _parameters.PrepareCapacity(parameterCount);
 
     for (csmInt32 i = 0; i < parameterCount; ++i)
     {
         Utils::Value& param = root[ExpressionKeyParameters][i];
         const CubismIdHandle parameterId = CubismFramework::GetIdManager()->GetId(param[ExpressionKeyId].GetRawString()); // パラメータID
-        const csmFloat32 value = static_cast<csmFloat32>(param[ExpressionKeyValue].ToFloat());   // 値
+        const csmFloat32 value = static_cast<csmFloat32>(param[ExpressionKeyValue].ToFloat());                            // 値
 
         // 計算方法の設定
         ExpressionBlendType blendType;
@@ -78,39 +110,10 @@ CubismExpressionMotion* CubismExpressionMotion::Create(const csmByte* buffer, cs
         item.BlendType   = blendType;
         item.Value       = value;
 
-        expression->_parameters.PushBack(item);
+        _parameters.PushBack(item);
     }
 
-    Utils::CubismJson::Delete(json); // JSONデータは不要になったら削除する
-
-    return expression;
-}
-
-void CubismExpressionMotion::DoUpdateParameters(CubismModel* model, csmFloat32 userTimeSeconds, csmFloat32 weight, CubismMotionQueueEntry* motionQueueEntry)
-{
-    for (csmUint32 i = 0; i < _parameters.GetSize(); ++i)
-    {
-        ExpressionParameter& parameter = _parameters[i];
-
-        switch (parameter.BlendType)
-        {
-        case ExpressionBlendType_Add: {
-            model->AddParameterValue(parameter.ParameterId, parameter.Value, weight);            // 相対変化 加算
-            break;
-        }
-        case ExpressionBlendType_Multiply: {
-            model->MultiplyParameterValue(parameter.ParameterId, parameter.Value, weight);       // 相対変化 乗算
-            break;
-        }
-        case ExpressionBlendType_Overwrite: {
-            model->SetParameterValue(parameter.ParameterId, parameter.Value, weight);            // 絶対変化 上書き
-            break;
-        }
-        default:
-            // 仕様にない値を設定したときは既に加算モードになっている
-            break;
-        }
-    }
+    Utils::CubismJson::Delete(json);// JSONデータは不要になったら削除する
 }
 
 }}}